requirements.txt
absl-py==0.8.1
astor==0.8.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.1.1
gast==0.2.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
numpy==1.17.4
oauthlib==3.1.0
opencv-python==4.1.1.26
opt-einsum==3.1.0
pandas==0.25.3
Pillow==6.2.1
protobuf==3.10.0
py4j==0.10.7
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyparsing==2.4.5
pyspark==2.4.4
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.3.2
six==1.13.0
tb-nightly==2.1.0a20191117
tensorflow==1.0.0
termcolor==1.1.0
tf-estimator-nightly==2.0.0.dev2019111709
tf-nightly==2.1.0.dev20191117
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
